+++ core/module/lang.cpp
@@ -131,7 +131,7 @@ void ObjeckLang::Execute(const std::wstring cmd_args)
 
   // initialize execution
   size_t* op_stack = new size_t[OP_STACK_SIZE];
-  long* stack_pos = new long;
+  size_t* stack_pos = new size_t;
   (*stack_pos) = 0;
 
   // execute
